Desktop layout + Disabling viewport zooming
Before responsive design was a thing, mobile devices only had a desktop
layout to work with. To cope with this, they zoomed out to fit the
entire desktop layout into the width of the screen, letting the user
interact with it by zooming in when necessary.
This default behavior will prevent mobile devices from using the mobile
layout, so it needs to be disabled.